+++ packages/browser/src/scenes/smartList/create-or-update/CreateSmartListScene.tsx
@@ -3,15 +3,15 @@ import React from 'react'
 import { SceneContainer } from '@/components/container'
 
 import CreateSmartListHeader from './CreateSmartListHeader'
-// import { CreateSmartListForm } from './form'
-import { SmartListQueryBuilder } from './queryBuilder'
+import { CreateSmartListForm } from './form'
+// import { SmartListQueryBuilder } from './queryBuilder'
 
 export default function CreateSmartListScene() {
 	return (
 		<>
 			<CreateSmartListHeader />
 			<SceneContainer>
-				<SmartListQueryBuilder />
+				<CreateSmartListForm />
 			</SceneContainer>
 		</>
 	)
