+++ packages/browser/src/scenes/smartList/create-or-update/form/CreateSmartListForm.tsx
@@ -1,38 +1,33 @@
-import { Form } from '@stump/components'
+import { zodResolver } from '@hookform/resolvers/zod'
+import { Form, Heading, Text } from '@stump/components'
 import React from 'react'
-import { useForm, UseFormReturn } from 'react-hook-form'
+import { useForm } from 'react-hook-form'
 
 import { ContentContainer } from '@/components/container'
 
+import { SmartListQueryBuilder } from '../queryBuilder'
 import AccessSettings from './AccessSettings'
 import BasicDetails from './BasicDetails'
-import FilterConfiguration from './FilterConfiguration'
-import { Schema } from './schema'
+import { schema, SmartListFormSchema } from './schema'
 
 export default function CreateSmartListForm() {
-	const form = useForm<Schema>({
+	const form = useForm<SmartListFormSchema>({
 		defaultValues: {
 			filters: {
 				groups: [
 					{
-						filters: [
-							// {
-							// 	name: 'test-1',
-							// },
-							// {
-							// 	name: 'test-2',
-							// },
-						],
-						joiner: 'and',
+						filters: [{ field: 'name', operation: 'any', source: 'book', value: ['boo', 'biz'] }],
+						joiner: 'or',
 					},
 				],
-				joiner: 'AND',
+				joiner: 'and',
 			},
 			visibility: 'PRIVATE',
 		},
+		resolver: zodResolver(schema),
 	})
 
-	const handleSubmit = async (data: Schema) => {
+	const handleSubmit = async (data: SmartListFormSchema) => {
 		console.debug('Create smart list', { data })
 	}
 
@@ -40,11 +35,19 @@ export default function CreateSmartListForm() {
 		<Form form={form} onSubmit={handleSubmit}>
 			<ContentContainer>
 				<BasicDetails />
-				<FilterConfiguration />
+
+				<div className="flex flex-col gap-y-6">
+					<div>
+						<Heading size="md">Filter configuration</Heading>
+						<Text variant="muted" size="sm">
+							The filters, y&apos;know
+						</Text>
+					</div>
+					<SmartListQueryBuilder />
+				</div>
+
 				<AccessSettings />
 			</ContentContainer>
 		</Form>
 	)
 }
-
-export type CreateSmartListForm = UseFormReturn<Schema>
