+++ packages/browser/src/scenes/smartList/create-or-update/queryBuilder/filterGroup/index.ts
@@ -1,0 +1,1 @@
+export { default as FilterGroup } from './FilterGroup'
