--- packages/browser/src/scenes/smartList/create-or-update/queryBuilder/filterValue/ListValue.tsx
@@ -1,4 +1,0 @@
-// TODO: number vs string list, use combobox!
-export default function ListValue() {
-	return <div></div>
-}
